requirements.txt
discord.py
aiohttp
Flask
Flask-Cors
Flask-Session
python-dotenv
itsdangerous
requests
ujson
PyYAML
pytz
colorama
loguru
watchdog
filelock
schedule
psutil
simplejson
orjson
